Run path detect if platform is not saved
This will force the installer to rerun the path auto detect for anyone upgrading from the previous Mod Manager who has a path saved.
While this is inconvenient for anyone who has BeatSaber in a different location, it should only happen on first launch.

Without this check, it would have prompted every single user to select which platform they are using. I'd rather not have that.